REFACTOR: move query builder to its own file
This commit moves everything from the query side that doesn't require a
database connection to query_builder.go
